+++ microscope/cameras/pvcam.py
@@ -144,6 +144,7 @@ import numpy as np
 import Pyro4
 
 import microscope
+import microscope._utils
 import microscope.abc
 
 
@@ -676,14 +677,16 @@ class md_frame(ctypes.Structure):
 
 if os.name == "nt":  # is windows
     if platform.architecture()[0] == "32bit":
-        _lib = ctypes.WinDLL("pvcam32")
+        _lib = microscope._utils.library_loader("pvcam32", ctypes.WinDLL)
     else:
-        _lib = ctypes.WinDLL("pvcam64")
+        _lib = microscope._utils.library_loader("pvcam64", ctypes.WinDLL)
 else:
-    _lib = ctypes.CDLL("pvcam.so")
+    _lib = microscope._utils.library_loader("pvcam.so")
+
 
 ### Functions ###
 STRING = ctypes.c_char_p
+
 
 # classes so that we do some magic and automatically add byrefs etc ... can classify outputs
 # (Nicked from PYME's Ixon wrapper.)
@@ -1691,6 +1694,7 @@ class PVCamera(
         _logger.info("Initializing %s", self._pv_name)
         self.handle = _cam_open(self._pv_name, OPEN_EXCLUSIVE)
         PVCamera.open_cameras.append(self.handle)
+
         # Set up event callbacks. Tried to use the resume callback to reinit camera
         # after power loss, but any attempt to close/reopen the camera or deinit the
         # DLL throws a Windows Error 0xE06D7363.
@@ -1717,7 +1721,7 @@ class PVCamera(
         )
         # Repopulate _params.
         self._params = {}
-        for (param_id, name) in _param_to_name.items():
+        for param_id, name in _param_to_name.items():
             try:
                 p = PVParam.factory(self, param_id)
             except: