+++ microscope/simulators/__init__.py
@@ -2,6 +2,7 @@
 
 ## Copyright (C) 2020 David Miguel Susano Pinto <carandraug@gmail.com>
 ## Copyright (C) 2020 Mick Phillips <mick.phillips@gmail.com>
+## Copyright (C) 2022 Ian Dobbie <ian.dobbie@gmail.com>
 ##
 ## This file is part of Microscope.
 ##
@@ -26,6 +27,7 @@ hardware behaviour.  They implement the different ABC.
 """
 
 import logging
+import math
 import random
 import time
 import typing
@@ -479,3 +481,96 @@ class SimulatedStage(microscope.abc.Stage):
     def move_to(self, position: typing.Mapping[str, float]) -> None:
         for name, pos in position.items():
             self.axes[name].move_to(pos)
+
+
+class SimulatedDigitalIO(microscope.abc.DigitalIO):
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+        self._cache = [None] * self._numLines
+        self.testinput = False
+        self.inputtime = time.time()
+
+    def set_IO_state(self, line: int, state: bool) -> None:
+        _logger.info("Line %d set IO state %s" % (line, str(state)))
+        self._IOMap[line] = state
+        if not state:
+            # this is an input so needs to have a definite value,
+            # default to False if not already set. If set leave alone
+            if self._cache[line] == None:
+                self._cache[line] = False
+
+    def get_IO_state(self, line: int) -> bool:
+        return self._IOMap[line]
+
+    def write_line(self, line: int, state: bool):
+        _logger.debug("Line %d set IO state %s" % (line, str(state)))
+        self._cache[line] = state
+
+    def read_line(self, line: int) -> bool:
+        _logger.debug("Line %d returns %s" % (line, str(self._cache[line])))
+        return self._cache[line]
+
+    def _do_shutdown(self) -> None:
+        pass
+
+    # functions required as we are DataDevice returning data to the server.
+    def _fetch_data(self):
+        if (time.time() - self.inputtime) > 5.0:
+            self.testinput = not self.testinput
+            self.inputtime = time.time()
+            _logger.debug("Line %d returns %s" % (3, self.testinput))
+            self._cache[3] = self.testinput
+            return (3, self.testinput)
+        return None
+
+    def abort(self):
+        pass
+
+    def _do_enable(self):
+        return True
+
+
+# DIO still to do:
+# raise exception if writing to a read line and vis-versa
+# raise exception if line <0 or line>num_lines
+# read all lines to return True,Flase if readable and None if an output
+#
+
+
+class SimulatedValueLogger(microscope.abc.ValueLogger):
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+        self._cache = [None] * self._numSensors
+        self.lastDataTime = time.time()
+
+    def initialize(self):
+        # init simulated sensors
+        for i in range(self._numSensors):
+            self._cache[i] = 20 + i
+
+    # functions required as we are DataDevice returning data to the server.
+    def _fetch_data(self):
+        if (time.time() - self.lastDataTime) > 5.0:
+            self.lastDataTime = time.time()
+            return self.getValues()
+        return None
+
+    def getValues(self):
+        for i in range(self._numSensors):
+            self._cache[i] = (
+                19.5
+                + i
+                + 5 * math.sin(self.lastDataTime / 100)
+                + random.random()
+            )
+            _logger.debug("Sensors %d returns %s" % (i, self._cache[i]))
+        return self._cache
+
+    def abort(self):
+        pass
+
+    def _do_enable(self):
+        return True
+
+    def _do_shutdown(self) -> None:
+        pass