+++ microscope/testsuite/test_device_server.py
@@ -84,7 +84,7 @@ def _patch_out_device_server_logs(func):
         return logging.NullHandler()
 
     no_file = unittest.mock.patch(
-        "microscope.device_server.RotatingFileHandler", null_logs
+        "microscope.device_server.FileHandler", null_logs
     )
     no_stream = unittest.mock.patch(
         "microscope.device_server.StreamHandler", null_logs
@@ -112,6 +112,7 @@ class BaseTestServeDevices(unittest.TestCase):
         options = microscope.device_server.DeviceServerOptions(
             config_fpath="",
             logging_level=logging.INFO,
+            logging_dir="",
         )
         self.p = multiprocessing.Process(
             target=microscope.device_server.serve_devices,
@@ -274,6 +275,7 @@ class TestServingFloatingDevicesWithWrongUID(BaseTestDeviceServer):
         microscope.device_server.DeviceServerOptions(
             config_fpath="",
             logging_level=logging.INFO,
+            logging_dir="",
         ),
         {"bar": "127.0.0.1"},
         {"bar": 8001},
@@ -307,6 +309,7 @@ class TestFunctionInDeviceDefinition(BaseTestDeviceServer):
         microscope.device_server.DeviceServerOptions(
             config_fpath="",
             logging_level=logging.INFO,
+            logging_dir="",
         ),
         {},
         {},